chore: more _ui and _utils
